### 2017-03-13 #206 @whedon
[PRE REVIEW]: Varistran: Anscombe's variance stabilizing transformation for RNA-seq gene expression data
**Submitting author:** @pfh (<a href="http://orcid.org/0000-0002-3980-268X">Paul Harrison</a>)
**Repository:** <a href="https://github.com/MonashBioinformaticsPlatform/varistran" target ="_blank">https://github.com/MonashBioinformaticsPlatform/varistran</a>
**Version:** v1.0.1
