### 2017-04-17 #206 @arfon
:wave: @noamross would this submission be suitable for rOpenSci?
