### 2017-04-17 #206 @noamross
No, as this is primarily a statistical method implementation, it's out of scope for us.
