### 2017-04-17 #206 @arfon
> No, as this is primarily a statistical method implementation, it's out of scope for us.

OK thanks for confirming.
