### 2017-07-30 6969f6f
Make changes suggested in JOSS review.
