+++ R/dendrogram.R
@@ -1,0 +1,182 @@
+
+# This code for drawing dendrograms has been adapted from ggdendro package.
+
+dendro_data <- function(model, type=c("rectangle","triangle"), ...) {
+  dhc <- as.dendrogram(model)
+  hcdata <- dendrogram_data(dhc, type=type, ...)
+  list(
+      segments = hcdata$segments,
+      labels = hcdata$labels,
+      class="hclust"
+  )
+}
+
+dendrogram_data <- function (x, type = c("rectangle", "triangle"), ...){ 
+
+    # Initialise variables that used to be in parameter list
+    leaflab <- "perpendicular"
+    center <- FALSE
+    xlab <- ""
+    ylab <- ""
+    horiz <- FALSE
+    #frame.plot <- FALSE
+    xaxt <- "n"
+    yaxt <- "s"
+    dLeaf <- NULL 
+    edge.root <- is.leaf(x) || !is.null(attr(x, "edgetext"))
+
+    type <- match.arg(type)
+    #leaflab <- match.arg(leaflab)
+    hgt <- attr(x, "height")
+    if (edge.root && is.logical(edge.root)) 
+        edge.root <- 0.0625 * if (is.leaf(x)) 1 else hgt
+    mem.x <- .memberDend(x)
+    yTop <- hgt + edge.root
+    if (center) {
+        x1 <- 0.5
+        x2 <- mem.x + 0.5
+    }
+    else {
+        x1 <- 1
+        x2 <- mem.x
+    }
+    xl. <- c(x1 - 1/2, x2 + 1/2)
+    yl. <- c(0, yTop)
+    if (edge.root) {
+        if (!is.null(et <- attr(x, "edgetext"))) {
+            my <- mean(hgt, yTop)
+        }
+    }
+    
+    gg.plotNode <- function (x1, x2, subtree, type, center, leaflab, dLeaf, horiz=FALSE, ddsegments=NULL, ddlabels=NULL) {
+        inner <- !is.leaf(subtree) && x1 != x2
+        yTop <- attr(subtree, "height")
+        bx <- plotNodeLimit(x1, x2, subtree, center)
+        xTop <- bx$x
+        hasP <- !is.null(nPar <- attr(subtree, "nodePar"))
+        Xtract <- function(nam, L, default, indx) rep(if (nam %in% 
+                                    names(L)) L[[nam]] else default, length.out = indx)[indx]
+        asTxt <- function(x){
+            if (is.character(x) || is.expression(x))
+                x else
+            if (is.null(x)) "" else as.character(x)
+        }
+        i <- if (inner || hasP) 
+                    1
+                else 2
+        lab.cex <- 1
+        if (is.leaf(subtree)) {
+            if (leaflab == "perpendicular") {
+                Y <- yTop - dLeaf * lab.cex
+                X <- xTop
+                srt <- 90
+                adj <- 1
+                nodeText <- asTxt(attr(subtree, "label"))
+                ddlabels <- rbind(ddlabels, data.frame(x=X, y=0, text=nodeText))
+            }
+        }
+        else if (inner) {
+            segmentsHV <- function(x0, y0, x1, y1) {
+                # *************************
+                data.frame(x0, y0, x1, y1) #AdV
+            }
+            for (k in seq_along(subtree)) {
+                child <- subtree[[k]]
+                yBot <- attr(child, "height")
+                if (getOption("verbose")) 
+                    cat("ch.", k, "@ h=", yBot, "; ")
+                if (is.null(yBot)) 
+                    yBot <- 0
+                xBot <- if (center) 
+                            mean(bx$limit[k:(k + 1)])
+                        else bx$limit[k] + .midDend(child)
+                if (type == "triangle") {
[…]
