+++ R/varistran.R
@@ -153,6 +153,17 @@ vst_methods <- list(
 #' @references Anscombe, F.J. (1948) "The transformation of Poisson, binomial,
 #' and negative-binomial data", Biometrika 35 (3-4): 246-254
 #'
+#' @examples
+#'
+#' # Generate some random data.
+#' means <- runif(100,min=0,max=1000)
+#' counts <- matrix(rnbinom(1000, size=1/0.01, mu=rep(means,10)), ncol=10)
+#'
+#' y <- varistran::vst(counts)
+#'
+#' # Information about the transformation
+#' varistran::vst_advice(y)
+#'
 #' @export
 vst <- function(x, method="anscombe.nb", lib.size=NULL, cpm=FALSE, dispersion=NULL, design=NULL) {
     x <- as.matrix(x)
