### 2017-07-30 5511252
Describe expected appearance of plot_stability.
