+++ R/plot.R
@@ -70,6 +70,8 @@
 #' Genes are partitioned evenly into "bins" bins by average expression level.
 #' Mean residual standard deviation is plotted against mean count.
 #'
+#' If the variance stabilizing transformation has been successful, this plot should be close to a horizontal line. However it is normal for the standard deviation to drop off for counts below 5.
+#'
 #' @param y Transformed counts matrix.
 #' @param x Optional, original counts matrix.
 #' @param design Matrix specifying a linear model with which to calculate
