### 2017-08-01 #257 @brainstorm
@roryk If you need some guidance regarding the remaining checkboxes, let me know... I actually need some guidance on the editorial side, @arfon. Is there something special I should do when Rory finishes the review? First editing role :)
