### 2017-08-02 #257 @arfon
> @arfon. Is there something special I should do when Rory finishes the review? First editing role :)

@brainstorm - once @roryk has finished his review ping me here and I can finalize the submission.
