### 2017-08-08 7946817
Remove 'Journal Article' text from reference -- it's obviously a journal article.
