+++ paper.bib
@@ -7,6 +7,5 @@
    number = {3/4},
    pages = {246-254},
    ISSN = {0006-3444},
-   year = {1948},
-   type = {Journal Article}
+   year = {1948}
 }
