### 2017-08-08 91105e0
Add support/contribute section and references to README.md
