### 2017-08-08 af803e2
Add support/contribute section and references to README.md
