### 2017-08-08 807a4ce
Add support/contribute section and references to README.md
