+++ README.md
@@ -104,6 +104,7 @@ varistran::shiny_report(counts=counts)
 
 * [Online demo](http://rnasystems.erc.monash.edu:3838/pfh/2015/demo-varistran)
 
+This also includes the [limma bioconductor package's](http://bioconductor.org/packages/release/bioc/html/limma.html) `plotMDS` MDS plot (Ritchie et al, 2015).
 
 ## Test suite
 
@@ -153,6 +154,8 @@ Frazee, Alyssa C., Ben Langmead and Jeffrey T. Leek. 2011. "ReCount: A multi-exp
 Love, Michael I., Wolfgang Huber and Simon Anders. 2014. "Moderated estimation of fold change and dispersion for RNA-seq data with DESeq2." *Genome Biology* 15 (12): 550. doi:10.1186/s13059-014-0550-8
 
 McCarthy, Davis J., Yunshun Chen and Gordon K. Smyth. 2012. "Differential expression analysis of multifactor RNA-Seq experiments with respect to biological variation." *Nucleic Acids Research* 40 (10): 4288-4297. doi:10.1093/nar/gks042
+
+Ritchie,  Matthew E., Belinda Phipson, Di Wu, Yifang Hu, Charity W. Law, Wei Shi and Gordon K. Smyth. 2015. "limma powers differential expression analyses for RNA-sequencing and microarray studies." *Nucleic Acids Research* 43 (7): e47. doi:10.1093/nar/gkv007
 
 Robinson, Mark D. and Alicia Oshlack. 2010. "A scaling normalization method for differential expression analysis of RNA-seq data." *Genome Biology* 11 (3): R25. doi:10.1186/gb-2010-11-3-r25
 
