### 2017-08-21 #257 @brainstorm
Thanks a lot @roryk, seems that they might have addressed that in their latests commits a few days ago?:

https://github.com/MonashBioinformaticsPlatform/varistran/commits/master
