### 2017-08-21 #257 @brainstorm
@arfon Clear to finalize/go then! Thanks much Rory! ;)
