### 2017-08-23 #257 @pfh
@arfon - Great! I've uploaded a tarball of the latest version to Zenodo, DOI is https://doi.org/10.5281/zenodo.846925
