### 2017-08-27 #257 @pfh
Great to have this reviewed, published and citable. It's a component in quite a lot of the work we do. Thanks everyone.
